tests: Add tests for cells
Tests for cells and how they interact with the planet.

You will have slightly more freedom on how to implement, but it is still
fairly straight forward. Once we get more into the groove of this things
will get more and more open end.

Also we will do code review directly on the master branch right now.
There are more proper workflows that involve multiple branches and
rewriting commits but I will keep it simple for now. The only reason
(for now) that I will ask you rewrite commits is if the message really
sucks. If that happens I will call you and walk you through it.